+++ .babelrc
@@ -10,7 +10,8 @@
     }]
   ],
   "plugins": [
-  "babel-plugin-transform-class-properties"
+  "babel-plugin-transform-class-properties",
+  "react-hot-loader/babel"
   ],
   "env": {
       "test": {
