+++ js/ClientApp.jsx
@@ -1,25 +1,15 @@
 import React from 'react';
 import { render } from 'react-dom';
-import { BrowserRouter, Route, Switch } from 'react-router-dom';
-import Landing from './Landing';
-import Search from './Search';
+import App from './App';
 
-const FourOFour = () => <h1>404</h1>;
+const renderApp = () => {
+    render(<App />, document.getElementById('app'));
+}
 
-const App = () => (
-  <div className="app">
-    <BrowserRouter>
-      <div className="app">
-        <Switch>
-          <Route exact path="/" component={Landing} />
-          <Route path="/search" component={Search} />
-          <Route component={FourOFour} />
-        </Switch>
-      </div>
+renderApp()
 
-    </BrowserRouter>
-
-  </div>
-);
-
-render(<App />, document.getElementById('app'));
+if (module.hot){
+    module.hot.accept('./App', () => {
+        renderApp()
+    })
+}
