+++ js/Landing.jsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 const Landing = () => (
   <div className="app">
     <div className="landing">
-      <h1>VideoQ</h1>
+      <h1>Browser Anythings</h1>
       <input type="text" placeholder="Search" />
       <Link to="/search"> Browser All</Link>
     </div>
