+++ webpack.config.js
@@ -1,14 +1,20 @@
 const path = require('path')
+const webpack = require('webpack')
 
 module.exports = {
     context: __dirname,
-    entry: "./js/ClientApp.jsx",
-    devtool: "source-map",
+    entry: ["./js/ClientApp.jsx",
+    'react-hot-loader/patch',
+    'webpack-dev-server/client?http://localhost:8080',
+    'webpack/hot/only-dev-server'],
+    devtool: "cheap-eval-source-map",
     output: {
         path: path.join(__dirname, 'public'),
-        filename: 'bundle.js'
+        filename: 'bundle.js',
+        publicPath: '/public/'
     },
     devServer: {
+        hot: true,
         publicPath: '/public/',
         historyApiFallback: true
     },
@@ -20,6 +26,10 @@ module.exports = {
         reasons: true,
         chunks: false
     },
+    plugins: [
+        new webpack.HotModuleReplacementPlugin(),
+        new webpack.NamedModulesPlugin()
+    ],
     module: {
         rules: [
             {
